touch up changes, fixed few logic holes, changed types, got main to compile

+++ includes/testWrapInput.h
@@ -7,7 +7,7 @@
 
 #include "./testGlobals.h"
 
-int __wrap_boundedInput(FILE* stream, char* dest, size_t dest_size);
+size_t __wrap_boundedInput(FILE* stream, char* dest, size_t dest_size);
 int __wrap_takeIndexInput(int max);
 int __wrap_takeDepthInput(void);
 char* __wrap_takeDirectoryInput(void);
